backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.1",
    "fastapi>=0.124.2",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.10",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "markupsafe<3.0.0",
    "polars>=1.36.1",
    "pypdf>=6.4.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]
